# RoyalFilms.Api/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RoyalFilms.Api",
    "version": "v1"
  },
  "paths": {
    "/api/Ciudad": {
      "get": {
        "tags": [
          "Ciudad"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Peliculas": {
      "get": {
        "tags": [
          "Peliculas"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Peliculas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PeliculaDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Peliculas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/PeliculaDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Peliculas"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "PeliculaDto": {
      "type": "object",
      "properties": {
        "idPelicula": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "genero": {
          "format": "int32",
          "type": "integer"
        },
        "poster": {
          "type": "string"
        },
        "trailer": {
          "type": "string"
        },
        "estado": {
          "type": "string"
        }
      }
    }
  }
}